requirements.txt
altgraph==0.17.4
azure-cognitiveservices-speech==1.40.0
packaging==24.1
pyinstaller-hooks-contrib==2024.8
setuptools==73.0.1
